+++ Sources/PeripheryKit/SourceGraph/Mutators/AssignOnlyPropertyReferenceEliminator.swift
@@ -6,6 +6,7 @@ final class AssignOnlyPropertyReferenceEliminator: SourceGraphMutator {
     private let configuration: Configuration
     private let retainAssignOnlyPropertyTypes: [String]
     private let defaultRetainedTypes = ["AnyCancellable", "Set<AnyCancellable>", "[AnyCancellable]", "NSKeyValueObservation"]
+    private let retainedAttributes = ["State", "Binding"]
 
     required init(graph: SourceGraph, configuration: Configuration) {
         self.graph = graph
@@ -20,6 +21,7 @@ final class AssignOnlyPropertyReferenceEliminator: SourceGraphMutator {
             guard let declaredType = property.declaredType,
                   !retainAssignOnlyPropertyTypes.contains(declaredType),
                   !graph.isRetained(property),
+                  property.attributes.isDisjoint(with: retainedAttributes),
                   !property.isComplexProperty,
                   // A protocol property can technically be assigned and never used when the protocol is used as an existential
                   // type, however communicating that succinctly would be very tricky, and most likely just lead to confusion.