+++ Sources/Shared/Configuration.swift
@@ -41,8 +41,8 @@ public final class Configuration {
     @Setting(key: "build_arguments", defaultValue: [])
     public var buildArguments: [String]
 
-    @Setting(key: "schemes_arguments", defaultValue: [])
-    public var schemesArguments: [String]
+    @Setting(key: "xcode_list_arguments", defaultValue: [])
+    public var xcodeListArguments: [String]
 
     @Setting(key: "retain_assign_only_property_types", defaultValue: [], valueSanitizer: PropertyTypeSanitizer.sanitize)
     public var retainAssignOnlyPropertyTypes: [String]
@@ -253,6 +253,10 @@ public final class Configuration {
             config[$buildArguments.key] = buildArguments
         }
 
+        if $xcodeListArguments.hasNonDefaultValue {
+            config[$xcodeListArguments.key] = xcodeListArguments
+        }
+
         if $relativeResults.hasNonDefaultValue {
             config[$relativeResults.key] = relativeResults
         }
@@ -350,8 +354,8 @@ public final class Configuration {
                 $cleanBuild.assign(value)
             case $buildArguments.key:
                 $buildArguments.assign(value)
-            case $schemesArguments.key:
-                $schemesArguments.assign(value)
+            case $xcodeListArguments.key:
+                $xcodeListArguments.assign(value)
             case $relativeResults.key:
                 $relativeResults.assign(value)
             case $retainCodableProperties.key:
@@ -398,6 +402,7 @@ public final class Configuration {
         $skipBuild.reset()
         $cleanBuild.reset()
         $buildArguments.reset()
+        $xcodeListArguments.reset()
         $relativeResults.reset()
         $retainCodableProperties.reset()
         $retainEncodableProperties.reset()