+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Unused import detection is now enabled by default.
 - Added the `--retain-encodable-properties` option to retain all properties on `Encodable` types only.
+- Added the `--xcode-list-arguments` option to pass additional arguments to `xcodebuild -list`.
 
 ##### Bug Fixes